chore: align code style

+++ apps/backend/src/workflows/company/steps/create-companies.ts
@@ -1,41 +1,25 @@
 import { createStep, StepResponse } from "@medusajs/framework/workflows-sdk";
-import { ContainerRegistrationKeys } from "@medusajs/utils";
-import {
-  ICompanyModuleService,
-  ModuleCreateCompany,
-  QueryCompany,
-} from "@starter/types";
+import { ICompanyModuleService, ModuleCreateCompany } from "@starter/types";
 import { COMPANY_MODULE } from "../../../modules/company";
 
 export const createCompaniesStep = createStep(
   "create-companies",
-  async (
-    input: ModuleCreateCompany,
-    { container }
-  ): Promise<StepResponse<QueryCompany, string>> => {
+  async (input: ModuleCreateCompany, { container }) => {
     const companyModuleService =
       container.resolve<ICompanyModuleService>(COMPANY_MODULE);
-    const createdCompany = await companyModuleService.createCompanies(input);
 
-    const query = container.resolve(ContainerRegistrationKeys.QUERY);
-
-    const {
-      data: [company],
-    } = await query.graph({
-      entity: "companies",
-      filters: {
-        id: createdCompany.id,
-      },
-      fields: ["*", "+employees"],
-    });
+    const company = await companyModuleService.createCompanies(input);
 
     return new StepResponse(company, company.id);
   },
   async (companyId: string, { container }) => {
+    if (!companyId) {
+      return;
+    }
+
     const companyModuleService =
       container.resolve<ICompanyModuleService>(COMPANY_MODULE);
-    await companyModuleService.deleteCompanies([companyId]);
 
-    return new StepResponse("Company deleted", companyId);
+    await companyModuleService.deleteCompanies([companyId]);
   }
 );
